backend/requirements.txt
# Federated Learning Server Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Machine Learning / Deep Learning
tensorflow>=2.15.0
keras>=3.0.0
joblib>=1.3.0

# Authentication & Database
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.3.0
python-dotenv>=1.0.0
